+++ include/gui/gui.h
@@ -29,7 +29,7 @@ typedef struct s_fps
 typedef struct s_gui
 {
 	bool		is_hidden;
-	double		hidden_ratio;
+	float		hidden_ratio;
 	t_gui_boxes	gui_boxes;
 	bool		color_picker_base_color_was_changed;
 	t_color		color_picker_base_color;
