+++ src/math/conversion.c
@@ -12,7 +12,7 @@
 
 #include <math.h>
 
-double	convert_degrees_to_radians(double degree)
+float	convert_degrees_to_radians(float degree)
 {
 	return (degree * M_PI / 180);
 }
