+++ src/math/vector/vector2_math.c
@@ -26,14 +26,14 @@ t_vector2	vector2_subtract(t_vector2 v1, t_vector2 v2)
 	return (v1);
 }
 
-t_vector2	vector2_multiply(t_vector2 v1, double value)
+t_vector2	vector2_multiply(t_vector2 v1, float value)
 {
 	v1.x *= value;
 	v1.y *= value;
 	return (v1);
 }
 
-t_vector2	vector2_divide(t_vector2 v1, double value)
+t_vector2	vector2_divide(t_vector2 v1, float value)
 {
 	v1.x /= value;
 	v1.y /= value;
