+++ src/math/vector/vector4_operations.c
@@ -14,7 +14,7 @@
 
 #include "math/vector.h"
 
-double	vector4_length_squared(t_vector4 vector)
+float	vector4_length_squared(t_vector4 vector)
 {
 	return (vector.x * vector.x
 		+ vector.y * vector.y
@@ -22,7 +22,7 @@ double	vector4_length_squared(t_vector4 vector)
 		+ vector.w * vector.w);
 }
 
-double	vector4_length(t_vector4 vector)
+float	vector4_length(t_vector4 vector)
 {
 	return (sqrt(vector4_length_squared(vector)));
 }
@@ -32,7 +32,7 @@ t_vector4	vector4_unit(t_vector4 vector)
 	return (vector4_divide(vector, vector4_length(vector)));
 }
 
-double	vector4_dot(t_vector4 v1, t_vector4 v4)
+float	vector4_dot(t_vector4 v1, t_vector4 v4)
 {
 	return (v1.x * v4.x + v1.y * v4.y + v1.z * v4.z + v1.w * v4.w);
 }
